moved component styles to their own folder
