+++ frontend/src/components/styles/Chat.css
